fix: 온보딩 약관 바텀시트 문제 해결
